+++ pytest_django/fixtures.py
@@ -28,7 +28,6 @@ def _django_db_setup(request,
     skip_if_no_django()
 
     from .compat import setup_databases, teardown_databases
-    from django.core import management
 
     # xdist
     if hasattr(request.config, 'slaveinput'):
@@ -38,10 +37,7 @@ def _django_db_setup(request,
 
     monkey_patch_creation_for_db_suffix(db_suffix)
 
-    # Disable south's syncdb command
-    commands = management.get_commands()
-    if commands['syncdb'] == 'south':
-        management._commands['syncdb'] = 'django.core'
+    _handle_south()
 
     with _django_cursor_wrapper:
         # Monkey patch Django's setup code to support database re-use
@@ -92,6 +88,40 @@ def _django_db_fixture_helper(transactional, request, _django_cursor_wrapper):
         case._pre_setup()
         request.addfinalizer(_django_cursor_wrapper.disable)
         request.addfinalizer(case._post_teardown)
+
+def _handle_south():
+    from django.conf import settings
+    if 'south' in settings.INSTALLED_APPS:
+        # Handle south.
+        from django.core import management
+
+        try:
+            # if `south` >= 0.7.1 we can use the test helper
+            from south.management.commands import patch_for_test_db_setup
+        except ImportError:
+            # if `south` < 0.7.1 make sure it's migrations are disabled
+            management.get_commands()
+            management._commands['syncdb'] = 'django.core'
+        else:
+            # Monkey-patch south.hacks.django_1_0.SkipFlushCommand to load
+            # initial data.
+            # Ref: http://south.aeracode.org/ticket/1395#comment:3
+            import south.hacks.django_1_0
+            from django.core.management.commands.flush import Command as FlushCommand
+            class SkipFlushCommand(FlushCommand):
+                def handle_noargs(self, **options):
+                    # Reinstall the initial_data fixture.
+                    from django.core.management import call_command
+                    # `load_initial_data` got introduces with Django 1.5.
+                    load_initial_data = options.get('load_initial_data', None)
+                    if load_initial_data or load_initial_data is None:
+                        # Reinstall the initial_data fixture.
+                        call_command('loaddata', 'initial_data', **options)
+                    # no-op to avoid calling flush
+                    return
+            south.hacks.django_1_0.SkipFlushCommand = SkipFlushCommand
+
+            patch_for_test_db_setup()
 
 ################ User visible fixtures ################