+++ tests/conftest.py
@@ -22,7 +22,7 @@ from .db_helpers import (create_empty_production_database, get_db_engine,
 @pytest.fixture(scope='function')
 def django_testdir(request, testdir, monkeypatch):
     if get_db_engine() in ('mysql', 'postgresql_psycopg2', 'sqlite3'):
-        # Django requires the production database to exists..
+        # Django requires the production database to exist.
         create_empty_production_database()
 
     if hasattr(request.node.cls, 'db_settings'):
@@ -75,3 +75,16 @@ def django_testdir(request, testdir, monkeypatch):
     testdir.create_app_file = create_app_file
 
     return testdir
+
+
+@pytest.fixture
+def django_testdir_initial(django_testdir):
+    """A django_testdir fixture which provides initial_data."""
+    django_testdir.makefile('.json', initial_data="""
+        [{
+            "pk": 1,
+            "model": "app.item",
+            "fields": { "name": "mark_initial_data" }
+        }]""")
+
+    return django_testdir